Project Organization Step-bar chart
