+++ Step Tracker/Managers/HealthKitManager.swift
